fix issue 46 > extracted WeekDay to its own file
