+++ src/features/calendar_views/month_view/WeekRow.js
@@ -1,18 +1,15 @@
-import React, { useRef, useContext } from "react";
+import React from "react";
 
 // PACKAGES
 import PropTypes from "prop-types";
 import exact from "prop-types-exact";
-import dayjs from "dayjs";
-
-// CONTEXT
-import { EventContext } from "../../../common/EventContext";
 
 // SCRIPTS
 import { sortEvents } from "../../../common/calendar";
 
 // COMPONENTS
 import EventSummary from "./EventSummary";
+import WeekDay from "./WeekDay";
 
 function WeekRow({ eventsPerDay, week, month, maxNumberOfEvents }) {
   return (
@@ -54,68 +51,6 @@ WeekDays.propTypes = exact({
   month: PropTypes.number.isRequired,
   eventsPerDay: PropTypes.object.isRequired,
   maxNumberOfEvents: PropTypes.number.isRequired,
-});
-
-const WeekDay = ({ day, month, events, maxNumberOfEvents }) => {
-  const remainingEventsNumber =
-    events && maxNumberOfEvents < events.length ? events.length - maxNumberOfEvents : undefined;
-
-  const monthDay = dayjs(day).date();
-  const titleRef = useRef();
-
-  const { displayModal } = useContext(EventContext);
-
-  const chooseStyle = () => {
-    const today = dayjs().startOf("day");
-    const currentDate = dayjs(day).startOf("day");
-    if (month !== currentDate.month() || (currentDate.isBefore(today) && month === today.month())) return "past";
-    if (currentDate.isSame(today)) return "today";
-    return "future";
-  };
-
-  const styles = {
-    past: "text-gray-500 bg-gray-400 bg-opacity-25 hover:text-gray-700 hover:bg-gray-500 hover:bg-opacity-25",
-    today: "text-blue-600 bg-blue-300 bg-opacity-25 hover:bg-blue-400 hover:bg-opacity-25",
-    future: "bg-white bg-opacity-25 text-gray-700 hover:bg-gray-300 hover:bg-opacity-25 hover:text-black",
-  };
-
-  const handleClick = () => displayModal(day);
-  const currentStyle = styles[chooseStyle()];
-
-  return (
-    <>
-      <div className={`${currentStyle} relative px-2 py-2 border-b border-b-500 cursor-pointer`} onClick={handleClick}>
-        <div className={`text-md font-semibold`} ref={titleRef}>
-          {monthDay}
-        </div>
-        <RemainingEventsNumber remainingEventsNumber={remainingEventsNumber} isShown={!!remainingEventsNumber} />
-      </div>
-    </>
-  );
-};
-
-WeekDay.propTypes = exact({
-  day: PropTypes.string.isRequired,
-  month: PropTypes.number.isRequired,
-  events: PropTypes.array,
-  maxNumberOfEvents: PropTypes.number.isRequired,
-});
-
-function RemainingEventsNumber({ remainingEventsNumber, isShown }) {
-  return (
-    <>
-      {isShown && (
-        <div className="absolute bottom-0 left-0 flex justify-end w-full">
-          <div className="text-xs font-semibold text-gray-600 pr-2">+ {remainingEventsNumber}</div>
-        </div>
-      )}
-    </>
-  );
-}
-
-RemainingEventsNumber.propTypes = exact({
-  remainingEventsNumber: PropTypes.number,
-  isShown: PropTypes.bool.isRequired,
 });
 
 function DailyEventList({ date, events, maxNumberOfEvents }) {
